NN/requirements.txt
# Зависимости для обучения нейросетей и вывода данных.
numpy==1.24.2
keras==2.11.0
tensorflow==2.11.1
nltk==3.8.1

# Зависимости API-сервера.
fastapi==0.92.0
uvicorn==0.20.0
pydantic==1.10.5
pyyaml==6.0
httptools==0.5.0
